// src/services/api/generated/model/linkDecisionRequestDecision.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Stock API
 * API de gestion de stock (produits, catégories, recettes, utilisateurs)
 * OpenAPI spec version: v0.0.1
 */

export type LinkDecisionRequestDecision = typeof LinkDecisionRequestDecision[keyof typeof LinkDecisionRequestDecision];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const LinkDecisionRequestDecision = {
  LINK: 'LINK',
  DECLINE: 'DECLINE',
} as const;