// src/services/api/generated/model/loginResponse.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Stock API
 * API de gestion de stock (produits, catégories, recettes, utilisateurs)
 * OpenAPI spec version: v0.0.1
 */
import type { UserSummary } from './userSummary';

export interface LoginResponse {
  accessToken?: string;
  refreshToken?: string;
  user?: UserSummary;
}